AlCalorieApp-Cloud/requirements.txt
# Core Streamlit dependencies
streamlit>=1.28.0,<2.0.0
python-dotenv>=1.0.0

# AI and ML libraries (YOLO11m only)
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0

# LangChain and LLM
langchain>=0.1.0,<1.0.0
langchain-community>=0.0.10,<1.0.0
langchain-groq>=0.0.1,<1.0.0

# Computer Vision (YOLO11m)
ultralytics>=8.0.0,<9.0.0
opencv-python-headless>=4.8.0,<5.0.0
Pillow>=10.0.0,<11.0.0

# Data processing and visualization
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.15.0,<6.0.0

# Web requests for enhanced agent
requests>=2.25.0,<3.0.0






